Update DB & Kontrak

+++ application/models/JabatanModel.php
@@ -7,7 +7,7 @@ class JabatanModel extends CI_Model
 {
     // Initial
     public $table = 'jabatan';
-    public $id = 'id';
+    public $id = 'id_jab';
     public $order = 'DESC';
 
     function __construct()
@@ -18,7 +18,7 @@ class JabatanModel extends CI_Model
     // datatables
     function json()
     {
-        $this->datatables->select('id, jabatan');
+        $this->datatables->select('id_jab, jabatan');
         $this->datatables->from('jabatan');
         $this->datatables->add_column(
             'action',
@@ -29,7 +29,7 @@ class JabatanModel extends CI_Model
                 form_open('master/Jabatan/delete/$1') .
                 form_button(['type' => 'submit', 'title' => 'Hapus', 'class' => 'btn btn-danger'], '<i class="fas fa-trash-alt"> </i>', 'onclick="javascript: return confirm(\'Yakin ingin hapus ?\')"') .
                 form_close() . '</div>',
-            'id'
+            'id_jab'
         );
         return $this->datatables->generate();
     }
@@ -44,7 +44,7 @@ class JabatanModel extends CI_Model
     // get total rows
     function total_rows($q = NULL)
     {
-        $this->db->like('id', $q);
+        $this->db->like('id_jab', $q);
         $this->db->or_like('jabatan', $q);
         $this->db->from($this->table);
         return $this->db->count_all_results();
@@ -54,7 +54,7 @@ class JabatanModel extends CI_Model
     function get_limit_data($limit, $start = 0, $q = NULL)
     {
         $this->db->order_by($this->id, $this->order);
-        $this->db->like('id', $q);
+        $this->db->like('id_jab', $q);
         $this->db->or_like('jabatan', $q);
         $this->db->limit($limit, $start);
         return $this->db->get($this->table)->result();
@@ -83,7 +83,12 @@ class JabatanModel extends CI_Model
     // delete data
     function delete($id)
     {
-        $this->db->where($this->id, $id);
-        $this->db->delete($this->table);
+        if (!$this->db->where('id_jabatan', $id)->from('pegawai')->count_all_results() > 0) {
+            $this->db->where($this->id, $id);
+            $this->db->delete($this->table);
+            return TRUE;
+        } else {
+            return FALSE;
+        }
     }
 }
